Quiz by cat (#78)
* create choose-cat page

* working on cat service

* add  quiz-music-service

* get all cats ok

* display cat ok

* display ok

* to be continued

* tbc

* correct long quiz

* correct long quiz errors

* correct last error

* Fix tests errors

---------

+++ server/src/main/java/com/wcs/server/repository/QuizRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface QuizRepository extends JpaRepository<Quiz, Integer> {
+public interface QuizRepository extends JpaRepository<Quiz, Long> {
     List<Quiz> findByCreatedBy(User user);
 
     List<Quiz> findQuizzesByCreatedBy(User user);
 
     @Query("SELECT id FROM Quiz")
-    List<Integer> findAllIds();
+    List<Long> findAllIds();
 }
